Tiny changes to README (update for v0.5)
Hey! I've updated the code in the README in a few places so it'd compile, plus made a few things clearer. For example, I've explicitly specified that the `Socket` object should be a part of the class (It confused me a little initially because I thought the `Socket` declaration and the `self.socket.channel` call were part of the same scope). Plus a few more changes so that the code would compile with the latest version of birdsong (`join()` now needs a `?`)